requirements.txt
fastapi[security]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
pydantic
pydantic-settings
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart